Add Gecko-specific endpoints for finding anonymous nodes.
XBL has the concept of anonymous nodes that are not returned by the
usual WebDriver element-finding methods. However there are two
Gecko-specific methods of finding them; either by getting all the
anonymous children of a reference element, or getting a single anonymous
child of a reference element with specified attribute values.

This commit adds two endpoints corresponding to those methods:

/session/{sessionId}/moz/xbl/{elementId}/anonymous_children

to return all anonymous children, and

/session/{sessionId}/moz/xbl/{elementId}/anonymous_by_attribute

to return an anonymous element with the given attribute value, provided
as a body of the form:

{
  "name": <attribute name>,
  "value": <attribute value>,
}